+++ api/org.integratedmodelling.klab.api/src/org/integratedmodelling/klab/api/knowledge/IObservable.java
@@ -380,63 +380,22 @@ public interface IObservable extends ISemantic, IResolvable {
 	 */
 	ICurrency getCurrency();
 
-//	/**
-//	 * The context type.
-//	 * 
-//	 * @return the context type
-//	 */
-//	IConcept getContext();
-//
-//	/**
-//	 * The inherent type.
-//	 * 
-//	 * @return the inherent type
-//	 */
-//	IConcept getInherentType();
-//
-//	/**
-//	 * The comparison type, if any, for observables that admit it - values,
-//	 * proportions and ratios. This is only certainly not null for ratios.
-//	 * 
-//	 * @return the inherent type
-//	 */
-//	IConcept getComparisonType();
-//
-//	/**
-//	 * The caused ('causing') type.
-//	 * 
-//	 * @return the caused type
-//	 */
-//	IConcept getCaused();
-//
-//	/**
-//	 * The causant ('caused by') type
-//	 * 
-//	 * @return the caused type
-//	 */
-//	IConcept getCausant();
-//
-//	/**
-//	 * The compresent ('with') type
-//	 * 
-//	 * @return the compresent type
-//	 */
-//	IConcept getCompresent();
-//
-//	/**
-//	 * The purpose ('for') type
-//	 * 
-//	 * @return the purpose type
-//	 */
-//	IConcept getPurpose();
-//
-//	/**
-//	 * The original type for a quality resulting from the application of a unary
-//	 * operator.
-//	 * 
-//	 * @return
-//	 */
-//	IConcept getDescribedType();
+	/**
+	 * The context type, direct or indirect, and revised according to the stated
+	 * inherency (will be reverted to null if the indirect context is X and the
+	 * concept is <this> of X). The inherency revision is only for observables and
+	 * does not affect the underlying semantics.
+	 * 
+	 * @return the context type
+	 */
+	IConcept getContext();
+
+	/**
+	 * The inherent type, direct or indirect.
+	 * 
+	 * @return the inherent type
+	 */
+	IConcept getInherent();
 
 	/**
 	 * If the observable was defined with an inline value (e.g. '10 as Concept'),