+++ klab.tools/src/main/java/org/integratedmodelling/klab/clitool/console/commands/model/Compatible.java
@@ -4,6 +4,7 @@ import java.util.List;
 import java.util.Set;
 
 import org.integratedmodelling.kim.api.IServiceCall;
+import org.integratedmodelling.klab.Concepts;
 import org.integratedmodelling.klab.Models;
 import org.integratedmodelling.klab.Observables;
 import org.integratedmodelling.klab.Reasoner;
@@ -17,7 +18,6 @@ import org.integratedmodelling.klab.utils.StringUtils;
 
 public class Compatible implements ICommand {
 
-	
 	/**
 	 * TODO add context parameter
 	 */
@@ -25,7 +25,14 @@ public class Compatible implements ICommand {
 	public Object execute(IServiceCall call, ISession session) throws KlabValidationException {
 
 		String ret = "";
+		String ctx = call.getParameters().get("context", String.class);
+
 		String declaration = StringUtils.join((List<?>) call.getParameters().get("arguments"), ' ').trim();
+
+		IConcept context = null;
+		if (ctx != null) {
+			context = Concepts.c(ctx);
+		}
 
 		IConcept concept = null;
 		if (declaration.startsWith("k:")) {
@@ -37,7 +44,7 @@ public class Compatible implements ICommand {
 		}
 
 		IObservable observable = Observable.promote(concept);
-		Set<Long> ids = Models.INSTANCE.getKbox().getCompatibleTypeIds(observable, null,
+		Set<Long> ids = Models.INSTANCE.getKbox().getCompatibleTypeIds(observable, context,
 				observable.getDescription().getResolutionMode());
 
 		for (long id : ids) {